refactor(native): split monitor runtime hotspots
