+++ native/rust/crates/ripdpi-monitor-engine/src/engine/runtime/stage.rs
@@ -1,52 +1,16 @@
+mod id;
+mod runner;
+
 use std::sync::Arc;
 use std::sync::atomic::AtomicBool;
 
 use rustls::client::danger::ServerCertVerifier;
 
 use super::plan::ExecutionPlan;
-use super::recording::{CollectedStageOutcome, record_collected_outcome};
+use super::recording::CollectedStageOutcome;
 use super::state::ExecutionRuntime;
 
-#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub(in crate::engine) enum ExecutionStageId {
-    Environment,
-    Dns,
-    Web,
-    Quic,
-    Tcp,
-    Service,
-    Circumvention,
-    Telegram,
-    Throughput,
-    DohJsonSurvey,
-    StrategyDnsBaseline,
-    StrategyTcpCandidates,
-    StrategyQuicCandidates,
-    StrategyConnectionConcurrency,
-    StrategyRecommendation,
-}
-
-impl ExecutionStageId {
-    pub(in crate::engine) const fn as_str(&self) -> &'static str {
-        match self {
-            Self::Environment => "environment",
-            Self::Dns => "dns",
-            Self::Web => "web",
-            Self::Quic => "quic",
-            Self::Tcp => "tcp",
-            Self::Service => "service",
-            Self::Circumvention => "circumvention",
-            Self::Telegram => "telegram",
-            Self::Throughput => "throughput",
-            Self::DohJsonSurvey => "doh_json_survey",
-            Self::StrategyDnsBaseline => "strategy_dns_baseline",
-            Self::StrategyTcpCandidates => "strategy_tcp_candidates",
-            Self::StrategyQuicCandidates => "strategy_quic_candidates",
-            Self::StrategyConnectionConcurrency => "strategy_connection_concurrency",
-            Self::StrategyRecommendation => "strategy_recommendation",
-        }
-    }
-}
+pub(in crate::engine) use id::ExecutionStageId;
 
 pub(in crate::engine) enum RunnerOutcome {
     Completed,
@@ -68,11 +32,7 @@ pub(in crate::engine) trait ExecutionStageRunner {
         runtime: &mut ExecutionRuntime,
         tls_verifier: Option<&Arc<dyn ServerCertVerifier>>,
     ) -> RunnerOutcome {
-        let deadline = runtime.scan_deadline();
-        let collected = ripdpi_diagnostics_contracts::util::with_scan_io_deadline(deadline, || {
-            self.run_collecting(plan, runtime.cancel_token(), tls_verifier)
-        });
-        record_collected_outcome(plan, runtime, collected)
+        runner::run(self, plan, runtime, tls_verifier)
     }
 
     /// Run without touching `runtime`, returning all steps collected independently.
